Create caesar.c
In the lecture that is associated with this code, David Malan made it seem (without explicitly stating it) that these codes should be written with just what was taught in the lecture.  I was unable to do that again.  I am not sure it is against the spirit of the program or not but I plan to continue using what makes sense to me and not restrict myself to the lectures.